abort releasing if version with same name already exists

+++ main.js
@@ -5,7 +5,7 @@ const child_process = require('child_process');
 // Make sure the tree is clean
 console.log('Making sure tree is clean...');
 if (child_process.execSync('git diff HEAD').length !== 0) {
-  console.log('Tree is dirty, aborting');
+  console.log('Tree is dirty, aborting...');
   process.exit(1);
 }
 
@@ -24,6 +24,10 @@ if (platform === 'mac') {
   const zipPath = process.argv[3];
   const zipName = zipPath.match('[^/]*$')[0];
   const zipDest = `mac/${zipName}`;
+  if (fs.existsSync(zipDest)) {
+    console.log(`'${zipDest}' already exists! Version name should be new. Aborting...`);
+    process.exit(1);
+  }
   console.log(`Moving '${zipPath}' to '${zipDest}'...`);
   fs.renameSync(zipPath, zipDest);
